+++ packages/client/package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "@reduxjs/toolkit": "1.8.6",
     "@types/react-redux": "7.1.24",
-    "@types/yup": "^0.32.0",
+    "@types/uuid": "8.3.4",
+    "@types/yup": "0.32.0",
     "classnames": "2.3.1",
     "dotenv": "16.0.2",
     "eslint-config-prettier": "8.5.0",
@@ -26,6 +27,7 @@
     "react-redux": "8.0.4",
     "react-router-dom": "6.4.2",
     "sass": "1.55.0",
+    "uuid": "9.0.0",
     "yup": "0.32.11"
   },
   "devDependencies": {
