+++ packages/client/src/components/ErrorsNotification/index.ts
@@ -1,0 +1,1 @@
+export { ErrorsNotification } from './ErrorsNotification';
