+++ packages/client/src/components/ErrorsNotification/types.ts
@@ -1,0 +1,5 @@
+export interface Props {
+  text: string
+  id: string
+  closeHandler: (id: string) => void
+}
