--- packages/client/src/components/PopupError/PopupError.module.scss
@@ -1,18 +1,0 @@
-@import 'src/index.module.scss';
-
-.error_block {
-  width: 100vw;
-  height: 50px;
-  background: #f35f63;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 1000;
-}
-
-.error_text {
-  margin: 0;
-  @include fontText(normal, 400, 18px, 50px);
-  color: #ffffff;
-  text-align: center;
-}
