--- packages/client/src/components/PopupError/index.ts
@@ -1,3 +1,0 @@
-import { PopupError } from './PopupError';
-
-export default PopupError;
