+++ packages/client/src/components/Preloader/Preloader.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './Preloader.module.scss';
 
-export const Preloader = (): JSX.Element => {
+export const Preloader = () => {
   return (
     <div className={styles.preloader_back}>
       <div className={styles.preloader} />
