+++ packages/client/src/layouts/BackgroundLayout/BackgroundLayout.tsx
@@ -1,15 +1,13 @@
 import styles from './BackgroundLayout.module.scss';
 
 type Props = {
-  children: JSX.Element
-}
+  children: JSX.Element;
+};
 
-const BackgroundLayout = ({ children }: Props) => {
+export const BackgroundLayout = ({ children }: Props) => {
   return (
     <div className={styles.container}>
       <div className={styles.wrapper}>{children}</div>
     </div>
   );
 };
-
-export default BackgroundLayout;
