+++ packages/client/src/layouts/BackgroundLayout/index.ts
@@ -1,1 +1,1 @@
-export { default } from './BackgroundLayout';
+export { BackgroundLayout } from './BackgroundLayout';
