--- packages/client/src/store/slices/types.ts
@@ -1,8 +1,0 @@
-import { IUser } from '../../modules/IUser';
-
-export interface AuthState {
-  checkAuth: boolean
-  auth: boolean
-  loading: boolean
-  user: IUser | null
-}
