+++ packages/client/src/store/types.ts
@@ -1,0 +1,15 @@
+import { IUser } from '../modules/IUser';
+
+export interface AuthState {
+  checkAuth: boolean
+  auth: boolean
+  loading: boolean
+  user: IUser | null
+}
+
+export interface ErrorState {
+  errorList: {
+    text: string
+    id: string
+  }[]
+}
