+++ packages/client/tsconfig.json
@@ -15,7 +15,8 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx"
+    "jsx": "react-jsx",
+    "typeRoots": ["node_modules/yup"]
   },
   "include": ["src"],
   "references": [{ "path": "./tsconfig.node.json" }],
